// plugin-or-cli/src/types/analytics.ts
export type ReviewerStats = {
  login: string;
  totalReviews: number;
  totalReviewComments: number;
  totalReviewCommentsAddressed: number;
  totalReviewCommentsIgnored: number;
  averageTimeSpentReviewing: number;
  totalTimeSpentReviewing: number;
  averageTimeBetweenReviews: number;
  averageReviewsPerDay: number;
  averageReviewsPerWeek: number;
  averageReviewsPerMonth: number;
  totalReviewsPerDay: {
    [date: string]: number;
  };
  totalReviewsPerWeek: {
    [date: string]: number;
  };
  totalReviewsPerMonth: {
    [date: string]: number;
  };
  totalReviewCommentsPerDay: {
    [date: string]: number;
  };
  totalReviewCommentsPerWeek: {
    [date: string]: number;
  };
  totalReviewCommentsPerMonth: {
    [date: string]: number;
  };
  requestToCompletionTimes: {
    [login: string]: number;
  };
  completionToAddressedTimes: {
    [login: string]: number;
  };
};